Move validator into Projects namespace
- Validator moves to app/validators/projects/identifier_validator.rb
  as Projects::IdentifierValidator. The path-style symbol on the
  validates declaration ('projects/identifier') resolves to it.
  Distinguishes domain-specific validators from globally reusable
  ones (UrlValidator, JsonValidator, etc).

- Spec moves to spec/validators/projects/identifier_validator_spec.rb.

- Trim comments that explained framework conventions (validator
  lookup, declaration order). Keep comments only where the WHY is
  non-obvious to a Rails-fluent reader.
